+++ notebooks/LUVOIR-A/ref_coro_system.jpg
